WIP: emit prospect Snowplow events
