+++ packages/prospectapi-common/events/snowplow.integration.ts
@@ -26,6 +26,21 @@ describe('snowplow', () => {
     expect(allEvents.bad).toEqual(0);
   });
 
+  it('should accept an event with a prospect', async () => {
+    const emitter = getEmitter();
+    const tracker = getTracker(emitter);
+
+    queueSnowplowEvent(tracker, 'prospect_reviewed', prospect);
+
+    // wait a sec * 3 because snowplow does internal queueing
+    await new Promise((resolve) => setTimeout(resolve, 3000));
+
+    const allEvents = await getAllSnowplowEvents();
+
+    expect(allEvents.total).toEqual(1);
+    expect(allEvents.bad).toEqual(0);
+  });
+
   it('should accept an event with a prospect with status reasons and no comment', async () => {
     const emitter = getEmitter();
     const tracker = getTracker(emitter);
