+++ servers/prospect-api/src/lib.spec.ts
@@ -7,7 +7,6 @@ import {
   getRandomizedSortedRankedProspects,
   isValidProspectType,
   parseReasonsCsv,
-  prospectToSnowplowProspect,
   standardizeLanguage,
   deDuplicateProspectUrls,
   sanitizeText,
@@ -380,72 +379,6 @@ describe('lib', () => {
 
       // length of returned array should be 4 after removing 2 duplicates
       expect(deDupedProspects.length).toEqual(4);
-    });
-  });
-
-  describe('prospectToSnowplowProspect', () => {
-    it('should add prospect details and authUserName to the snowplow entity', () => {
-      const ps: Prospect[] = makeProspects(1);
-
-      const snowplowProspect = prospectToSnowplowProspect(ps[0], 'LDAP|User');
-
-      expect(snowplowProspect.prospect_id).toEqual(ps[0].prospectId);
-      expect(snowplowProspect.scheduled_surface_id).toEqual(
-        ps[0].scheduledSurfaceGuid,
-      );
-      expect(snowplowProspect.prospect_source).toEqual(ps[0].prospectType);
-      expect(snowplowProspect.topic).toEqual(ps[0].topic);
-      expect(snowplowProspect.url).toEqual(ps[0].url);
-      expect(snowplowProspect.created_at).toEqual(ps[0].createdAt);
-      expect(snowplowProspect.reviewed_by).toEqual('LDAP|User');
-    });
-
-    it('should add status reasons to the snowplow entity if present', () => {
-      const ps: Prospect[] = makeProspects(1);
-
-      const snowplowProspect = prospectToSnowplowProspect(
-        ps[0],
-        'LDAP|User',
-        ['PUBLISHER', 'TOPIC'],
-        'allow me to explain...',
-      );
-
-      expect(snowplowProspect.prospect_id).toEqual(ps[0].prospectId);
-      expect(snowplowProspect.scheduled_surface_id).toEqual(
-        ps[0].scheduledSurfaceGuid,
-      );
-      expect(snowplowProspect.prospect_source).toEqual(ps[0].prospectType);
-      expect(snowplowProspect.topic).toEqual(ps[0].topic);
-      expect(snowplowProspect.url).toEqual(ps[0].url);
-      expect(snowplowProspect.created_at).toEqual(ps[0].createdAt);
-      expect(snowplowProspect.status_reasons).toEqual(['PUBLISHER', 'TOPIC']);
-      expect(snowplowProspect.status_reason_comment).toEqual(
-        'allow me to explain...',
-      );
-      expect(snowplowProspect.reviewed_by).toEqual('LDAP|User');
-    });
-
-    it('should skip adding status reasons to the snowplow entity if null', () => {
-      const ps: Prospect[] = makeProspects(1);
-
-      const snowplowProspect = prospectToSnowplowProspect(
-        ps[0],
-        'LDAP|User',
-        null,
-        null,
-      );
-
-      expect(snowplowProspect.prospect_id).toEqual(ps[0].prospectId);
-      expect(snowplowProspect.scheduled_surface_id).toEqual(
-        ps[0].scheduledSurfaceGuid,
-      );
-      expect(snowplowProspect.prospect_source).toEqual(ps[0].prospectType);
-      expect(snowplowProspect.topic).toEqual(ps[0].topic);
-      expect(snowplowProspect.url).toEqual(ps[0].url);
-      expect(snowplowProspect.created_at).toEqual(ps[0].createdAt);
-      expect(snowplowProspect.status_reasons).toBeUndefined;
-      expect(snowplowProspect.status_reason_comment).toBeUndefined;
-      expect(snowplowProspect.reviewed_by).toEqual('LDAP|User');
     });
   });
